requirements.txt
pandas
streamlit
google-cloud-vision
google-cloud
